Initial setup for SlideLayout work
