+++ src/components/slide-layout.tsx
@@ -1,0 +1,26 @@
+import Slide, { SlideProps } from './slide/slide';
+import { Box, FlexBox } from './layout-primitives';
+import { ReactNode } from 'react';
+
+const Center = ({ children, ...rest }: SlideProps) => (
+  <Slide {...rest}>
+    <FlexBox justifyContent="center" alignItems="center" height="100%">
+      <Box>{children}</Box>
+    </FlexBox>
+  </Slide>
+);
+
+export const TwoColumn = ({
+  left,
+  right,
+  ...rest
+}: Omit<SlideProps, 'children'> & { left: ReactNode; right: ReactNode }) => (
+  <Slide {...rest}>
+    <FlexBox flexDirection="row" alignItems="start" flex={1}>
+      <Box width="100%">{left}</Box>
+      <Box width="100%">{right}</Box>
+    </FlexBox>
+  </Slide>
+);
+
+export default { Center, TwoColumn };
